+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added Oracle, Microsoft SQLServer, Redshift in DB templates (#72)
 * Auto save form fields when switching connection labels (#71)
+* Add configuration settings to hide `Format SQL` and `Deploy Notebook` buttons
 
 ## 0.2.1